[pdsl] Temporarily remove kcov CI
The problem is that on TravisCI kcov is errornously missing even though we basically reused the travis CI yml specs from the stevia project where everything works just fine. Maybe it has something to do with caching and it only works for stevia since it has a cached kcov installation. We should investigate more.