[lang/ir] implement lint to ensure there are no __ink_ prefixed identifiers
Identifiers starting with __ink_ may only be used by the ink! codegen.
So far this has been only implemented on TokenTree or TokenStream level and not on parsed syn abstractions leading to arcane errors. E.g. it was possible to catch errors if an __ink_ substring appeared in a string literal before.