add ink_lang::ContractRef utility codegen type
This is going to be used as a base for all generated cross-calling contract types.
The reason is that it is hard to type-safely reference some generated type but it is easy to refer to a generic external one when the generic parameter is already given.
We will use the ink! storage struct as the generic parameter to solve the ink-as-dependency storage struct type generation using this utility type.