allow parameterizing ink::trait_definition macro
Also change the way how verify_hash is computed.
The verify_hash is no longer depending on the structure of the ink! trait definition. It only depends on the identifier as well as the optionally given namespace.
This allows more fine grained control over the resulting selector of an ink! trait definition.