add ink! trait message local ID utility method
This is required in order to add compile time information about ink! trait messages that can be queried by ink! trait definition implementers (smart contracts) later on.
The querying is required so that the smart contract is able to verify selectors and payable properties of its implemented ink! trait messages.
Also the user might safety guard against changes to the selector or payable property that can be validated by this means.