Fix ordering of message ids if the trait is implemented before the inherent section (#1235)
* Fixed the ordering of ids. Before it was sorted:
inherent_ids
traits_ids

Now the ordering is not changed. It is important for dispatching logic and wildcard. Dispatching takes the global index of the method but if ordering is changed, it will use wrong id. In the case of wildcard it breaks it at all if trait was implemented before inherent section.

* Make lcippy happy

* Update UI test to check that ids have right ordering with trait sections

* Move the method back to inherent section to fix ink waterfall

* Apply suggestions from code review

Co-authored-by: Michael Müller <mich@elmueller.net>

* Remove trait

Co-authored-by: Michael Müller <mich@elmueller.net>