Add ability to configure static buffer capacity
Currently, `StaticBuffer` in env/src/engine/on_chain/buffer.sh has a
static (constant) size of 16 kB. However, there are cases which might
require more space. One such case is passing in and getting out bigger
buffers to/from a chain extension. As chain extensions can be arbitrary,
it makes sense to have the ability to change the size of the buffer.

This commit adds the ability to configure the static buffer capacity at
compile time through the `ENV_ENGINE_STATIC_BUFFER_CAPACITY` environment
variable. For example, to double it to 32 kB, one might do the following
when compiling a smart contract:
```
ENV_ENGINE_STATIC_BUFFER_CAPACITY=32768 cargo +nightly contract build
```

In order to implement, we introduce the `ink_build_script_utils` crate
that has functions to read the value from the environment variable and
convert it to a Rust constant in a generated file at build-time. Then,
we utilize a build.rs script at the `env` crate level.

Documentation about the `ENV_ENGINE_STATIC_BUFFER_CAPACITY` environment
variable can be added in a separate commit to the `ink-docs` repo.