Update the release checklist
This simplies the release checklist a bit by using both `cargo-unleash`
and `cargo-release`. `cargo-release` has been way more reliable for
publishing the ink! workspace than `cargo-unleash`, but we still use
`cargo-unleash` to clean up our `dev-dependencies`.

In the future it would be cool to automate more of the release process,
such as tag creation and publishing, through `cargo-release`.