Update the release checklist (#1366)
* Update the release checklist

This simplies the release checklist a bit by using both `cargo-unleash`
and `cargo-release`. `cargo-release` has been way more reliable for
publishing the ink! workspace than `cargo-unleash`, but we still use
`cargo-unleash` to clean up our `dev-dependencies`.

In the future it would be cool to automate more of the release process,
such as tag creation and publishing, through `cargo-release`.

* Remove steps related to `dev-dependencies`

With #1367 these shouldn't be required anymore.

* Add missing punctuation

Co-authored-by: Michael Müller <mich@elmueller.net>

Co-authored-by: Michael Müller <mich@elmueller.net>