Generate checked and unchecked versions of contract messages
We're now able to specify if we want to use an unchecked or checked
version of the message functions in our code. This means that existing
code, such as in tests, doesn't need to worry about the new wrapped
`Result` type.

At the same time, as far as external callers are concerned they are
always calling the checked version of the messages.